Updates ConsoleUI

+++ src/Domain/SensorValidator.cpp
@@ -146,7 +146,7 @@ bool SensorValidator::isUserReliable( User& user) {
     vector<Sensor> sensors = user.getAssociatedSensors();
     int nbSensorsFalse = 0;
     if (sensors.empty()) return false;
-    for (int i = 0; i < sensors.size(); i++) {
+    for (size_t i = 0; i < sensors.size(); i++) {
         Sensor sensor = sensors[i];
         bool isValid = isValidSensor(sensor);
         if (!isValid) {
